moved get_remaining_file_list_for_args
