+++ sciencebeam/pipeline_runners/local_pipeline_runner.py
@@ -5,25 +5,12 @@ import logging
 import concurrent.futures
 from functools import partial
 from mimetypes import guess_type
-from typing import Callable, List
+from typing import Callable
 
 from sciencebeam_utils.beam_utils.io import (
     read_all_from_path,
     save_file_content
 )
-
-from sciencebeam_utils.beam_utils.files import find_matching_filenames_with_limit
-
-from sciencebeam_utils.utils.file_path import (
-    join_if_relative_path,
-    get_output_file
-)
-
-from sciencebeam_utils.utils.file_list import (
-    load_file_list
-)
-
-from sciencebeam_utils.tools.check_file_list import map_file_list_to_file_exists
 
 from sciencebeam.utils.formatting import format_size
 
@@ -44,6 +31,8 @@ from sciencebeam.pipeline_runners.pipeline_runner_utils import (
     add_batch_args,
     process_batch_args,
     encode_if_text_type,
+    get_output_file_for_source_file_fn,
+    get_remaining_file_list_for_args,
     DataProps
 )
 
@@ -77,26 +66,6 @@ def parse_args(pipeline, config, argv=None):
     return args
 
 
-def get_file_list_for_args(args: argparse.Namespace):
-    if args.source_file_list:
-        file_list_path = join_if_relative_path(args.base_data_path, args.source_file_list)
-        return load_file_list(
-            file_list_path, column=args.source_file_column, limit=args.limit
-        )
-    return list(find_matching_filenames_with_limit(
-        join_if_relative_path(args.base_data_path, args.source_path), limit=args.limit
-    ))
-
-
-def get_output_file_for_source_file_fn(args):
-    return partial(
-        get_output_file,
-        source_base_path=args.base_data_path,
-        output_base_path=args.output_path,
-        output_file_suffix=args.output_suffix
-    )
-
-
 def process_file(
         file_url: str, simple_runner: SimplePipelineRunner,
         get_output_file_for_source_url: Callable[[str], str]):
@@ -114,43 +83,16 @@ def process_file(
     LOGGER.info('saved output to: %s (%s)', output_file_url, format_size(len(output_content)))
 
 
-def get_file_list_without_output_file(
-        file_list: List[str],
-        get_output_file_for_source_url: Callable[[str], str]) -> List[str]:
-    output_file_exists_list = map_file_list_to_file_exists([
-        get_output_file_for_source_url(file_url)
-        for file_url in file_list
-    ])
-    LOGGER.debug('output_file_exists_list: %s', output_file_exists_list)
-    return [
-        file_url
-        for file_url, output_file_exists in zip(file_list, output_file_exists_list)
-        if not output_file_exists
-    ]
-
-
 def run(args, config, pipeline: Pipeline):
     LOGGER.info('args: %s', args)
-    file_list = get_file_list_for_args(args)
+    file_list = get_remaining_file_list_for_args(args)
     LOGGER.debug('file_list: %s', file_list)
-
-    if not file_list:
-        LOGGER.warning('no files found')
-        return
-
-    simple_runner = SimplePipelineRunner(pipeline.get_steps(config, args))
-
-    LOGGER.info('total number of files: %d', len(file_list))
-    if args.resume:
-        file_list = get_file_list_without_output_file(
-            file_list,
-            get_output_file_for_source_url=get_output_file_for_source_file_fn(args)
-        )
-        LOGGER.info('remaining number of files: %d', len(file_list))
 
     if not file_list:
         LOGGER.info('no files to process')
         return
+
+    simple_runner = SimplePipelineRunner(pipeline.get_steps(config, args))
 
     process_file_url = partial(
         process_file,
