+++ sciencebeam/pipeline_runners/pipeline_runner_utils.py
@@ -1,7 +1,26 @@
+import argparse
+import logging
 import os
-from argparse import ArgumentParser
+from functools import partial
+from typing import Callable, List
 
 from six import text_type
+
+from sciencebeam_utils.beam_utils.files import find_matching_filenames_with_limit
+
+from sciencebeam_utils.utils.file_path import (
+    join_if_relative_path,
+    get_output_file
+)
+
+from sciencebeam_utils.utils.file_list import (
+    load_file_list
+)
+
+from sciencebeam_utils.tools.check_file_list import map_file_list_to_file_exists
+
+
+LOGGER = logging.getLogger(__name__)
 
 
 class DataProps:
@@ -11,7 +30,7 @@ class DataProps:
     TYPE = 'type'
 
 
-def add_batch_args(parser: ArgumentParser):
+def add_batch_args(parser: argparse.ArgumentParser):
     parser.add_argument(
         '--data-path', type=str, required=True,
         help='base data path'
@@ -72,3 +91,56 @@ def process_batch_args(args):
 
 def encode_if_text_type(data):
     return data.encode('utf-8') if isinstance(data, text_type) else data
+
+
+def get_file_list_for_args(args: argparse.Namespace):
+    if args.source_file_list:
+        file_list_path = join_if_relative_path(args.base_data_path, args.source_file_list)
+        return load_file_list(
+            file_list_path, column=args.source_file_column, limit=args.limit
+        )
+    return list(find_matching_filenames_with_limit(
+        join_if_relative_path(args.base_data_path, args.source_path), limit=args.limit
+    ))
+
+
+def get_file_list_without_output_file(
+        file_list: List[str],
+        get_output_file_for_source_url: Callable[[str], str]) -> List[str]:
+    output_file_exists_list = map_file_list_to_file_exists([
+        get_output_file_for_source_url(file_url)
+        for file_url in file_list
+    ])
+    LOGGER.debug('output_file_exists_list: %s', output_file_exists_list)
+    return [
+        file_url
+        for file_url, output_file_exists in zip(file_list, output_file_exists_list)
+        if not output_file_exists
+    ]
+
+
+def get_output_file_for_source_file_fn(args):
+    return partial(
+        get_output_file,
+        source_base_path=args.base_data_path,
+        output_base_path=args.output_path,
+        output_file_suffix=args.output_suffix
+    )
+
+
+def get_remaining_file_list_for_args(args: argparse.Namespace):
+    file_list = get_file_list_for_args(args)
+    LOGGER.debug('file_list: %s', file_list)
+
+    if not file_list:
+        LOGGER.warning('no files found')
+        return
+
+    LOGGER.info('total number of files: %d', len(file_list))
+    if args.resume:
+        file_list = get_file_list_without_output_file(
+            file_list,
+            get_output_file_for_source_url=get_output_file_for_source_file_fn(args)
+        )
+        LOGGER.info('remaining number of files: %d', len(file_list))
+    return file_list
